+++ changelog.md
@@ -4,7 +4,7 @@
 
 ### Patch
 
-- Updated dev dependencies.
+- Updated dependencies.
 - Alphabetically sort exports in index files.
 
 ## 1.0.1