+++ changelog.md
@@ -1,6 +1,6 @@
 # device-agnostic-ui changelog
 
-## Next
+## 2.0.0
 
 ### Major